FROMLIST: drm/print: Add drm_debug_category_printer
This patch adds a new printer which will select the appropriate output
for a given debug category. Currently there is only one output target,
which is syslog. However in the future we'll have tracefs and it will be
useful to print to syslog, tracefs, or both. Drivers just need to create
the printer for the appropriate category and the printer will decide
where to send the output.

Signed-off-by: Sean Paul <seanpaul@chromium.org>

Changes in v5:
-Added to the set

(am from https://patchwork.freedesktop.org/patch/msgid/20200608210505.48519-11-sean@poorly.run)

BUG=b:160248414
TEST=Tested on hatch, observed traces in /sys/kernel/debug/tracing/
Change-Id: I889605edfc67f09dd3fbed6644f82a73f93a4316

Change-Id: Id651969cefcc38ecefaecc00c35187a03a0403cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2276797
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Sean Paul <seanpaul@chromium.org>
Commit-Queue: Sean Paul <seanpaul@chromium.org>