FROMGIT: spi: spi-geni-qcom: No need for irqsave variant of spinlock calls
The driver locks its locks in two places.

In the first usage of the lock the function doing the locking already
has a sleeping call and thus we know we can't be called from interrupt
context.  That means we can use the "spin_lock_irq" variant of the
function.

In the second usage of the lock the function is the interrupt handler
and we know interrupt handlers are called with interrupts disabled.
That means we can use the "spin_lock" variant of the function.

This patch is expected to be a no-op and is just a cleanup / slight
optimization.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Link: https://lore.kernel.org/r/20200616034044.v3.1.Ic50cccdf27d42420a63485082f8b5bf86ed1a2b6@changeid
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 539afdf969d6ad7ded543d9abf14596aec411fe9
 git://git.kernel.org/pub/scm/linux/kernel/git/broonie/spi.git for-next)

BUG=b:122480751
TEST=Stress SPI

Change-Id: Ic50cccdf27d42420a63485082f8b5bf86ed1a2b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2253892
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>
(cherry picked from commit 5641995367233a61c11c9d37fc52c31cc48dd7e8)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2256762
Reviewed-by: Philip Chen <philipchen@chromium.org>
Commit-Queue: Philip Chen <philipchen@chromium.org>
Tested-by: Philip Chen <philipchen@chromium.org>