CHROMIUM: usb: typec: intel_pmc_mux: Check the port status before connect
The PMC microcontroller that we use for configuration, does
not supply any status information back. For port status we
need to talk to another controller on the board called IOM
(I/O manager).

By checking the port status before configuring the muxes, we
can make sure that we do not reconfigure the port after
bootup when the system firmware (for example BIOS) has
already configured it.

Using the status information also to check if DisplayPort
HPD is still asserted when the cable plug is disconnected,
and clearing it if it is.

BUG=b:156016218
TEST=Ensure USB mass storage device is detected properly, when
volteer is booted with the mass storage device connected.
TOREVERT: Once the upstream patch lands in maintainer's branch,
this patch should be reverted.
Upstream patch under review is here
https://patchwork.kernel.org/patch/11666449

Change-Id: I30cf869796bed021eca84246fccd2e183be2cde6
Signed-off-by: Heikki Krogerus <heikki.krogerus@linux.intel.com>
Signed-off-by: Rajmohan Mani <rajmohan.mani@intel.com>
Signed-off-by: Azhar Shaikh <azhar.shaikh@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2317035
Tested-by: Chiranjeevi Rapolu <chiranjeevi.rapolu@intel.corp-partner.google.com>
Reviewed-by: Prashant Malani <pmalani@chromium.org>