CHROMIUM: thunderbolt: Use bit 31 to check if firmware CM is running in Tiger Lake
In Tiger Lake the firmware CM is always enabled (so bit 0 is always set)
but it may be in "pass through" mode which means it requires software CM
instead. This can be determined by checking bit 31 instead.

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>

BUG=b:160191015
TEST=Observed no regressions with S0iX without an attached device.

TOREVERT=Once the upstream version of this patch lands in maintainer
tree, this patch should be reverted.

Signed-off-by: Casey Bowman <casey.g.bowman@intel.com>
Change-Id: I0d4b846b3636709099b27918027fd55eceb77ae1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2275606
Tested-by: Dana Alkattan <dana.alkattan@intel.com>
Reviewed-by: Prashant Malani <pmalani@chromium.org>
Commit-Queue: Prashant Malani <pmalani@chromium.org>