BACKPORT: ASoC: Intel: SOF: merge COMETLAKE_LP and COMETLAKE_H
We already have two configurations for CometLake, and a third one
coming. On other platforms, we used a single Kconfig option, so we
should follow the same trend by merging the two cases in a backwards
compatible way.

The backwards compatibility is handled by overloading the COMETLAKE_LP
kconfig as COMETLAKE. In practice we've never seen a case where
COMETLAKE_H is not selected along with COMETLAKE_LP, so keeping one
of the two is enough.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Reviewed-by: Guennadi Liakhovetski <guennadi.liakhovetski@linux.intel.com>
Reviewed-by: Kai Vehmanen <kai.vehmanen@linux.intel.com>
Link: https://lore.kernel.org/r/20200617164755.18104-2-pierre-louis.bossart@linux.intel.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 4228668eb936357657046b486207b167caea5175)
Signed-off-by: Chintan Patel <chintan.m.patel@intel.com>

 Conflicts:
	sound/soc/sof/intel/Kconfig
(cmpatel) Minor conflicts

BUG=b:162008784
TEST=Test audio on volteer.

Change-Id: Ifef1edc9c54b9f3aac9e620ec4712d8b0b8ccc9c
Signed-off-by: Mike Mason <michael.w.mason@intel.corp-partner.google.com>