Revert "FROMGIT: elants_i2c - Report resolution information for touch major"
This reverts commit cb56746ac528e3d0c3f2d296bb7a2fdd1a87d1a8.

This commit is causing b/158477364 on Dru and part of b/159195016 on
Waddledoo, so it seems that there are multiple ELAN touchscreens for
which the resolution of 1 unit per mm is incorrect. Reverting on all
Kernel versions seems like a prudent step until we have a more universal
way to set the resolution.

BUG=b:160021854
TEST=none

Signed-off-by: Harry Cutts <hcutts@chromium.org>
Change-Id: I71dc904ba9652b0204fe463e55d87ab7a8488aad
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2252779
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
(cherry picked from commit 2855771fa1613549becbcb443fab480a3f371879)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2274899
Commit-Queue: Dmitry Torokhov <dtor@chromium.org>