FROMGIT: thunderbolt: No need to warn in TB_CFG_ERROR_INVALID_CONFIG_SPACE
This may be returned for example when accessing some of the vendor
specific capabilities. It is not fatal by any means so instead of WARN()
just log it as debug level. The caller gets error back anyway and is
expected to handle it accordingly.

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
Reviewed-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit fa1653d99cc8296628e4fa949afad6f87889d14b
https://git.kernel.org/pub/scm/linux/kernel/git/westeri/thunderbolt.git next)

BUG=b:161731926
TEST=ls /sys/kernel/debug/thunderbolt/
TEST=Confirm TBT devices exist when hotplugged.

Signed-off-by: Casey Bowman <casey.g.bowman@intel.com>
Change-Id: Ib7f9f3d17a4e10d64f08823529a97ab339895e95
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2381315
Tested-by: Madhulika Dolke <madhulika.dolke@intel.com>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Prashant Malani <pmalani@chromium.org>