FROMGIT: net: ipa: reduce aggregation time limit
Halve the time limit used when aggregation is enabled on an RX
endpoint, to half a millisecond.

Use DIV_ROUND_CLOSEST() to compute the value that represents the
time period, to get better accuracy in the event the time limit is
not an even multiple of the granularity.

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 1d86652b13e8404ee42b6575d97ed228e92547d6
 git://git.kernel.org/pub/scm/linux/kernel/git/netdev/net-next.git master)

BUG=b:74020136
TEST=Booted trogdor, modem call functional

Signed-off-by: Sujit Kautkar <sujitka@chromium.org>
Change-Id: I0e4a6e02d931eae3ad618998dbeb16db99c1b3a8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2283974
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Evan Green <evgreen@chromium.org>