BACKPORT: FROMGIT: ALSA: hda/hdmi: Add Intel silent stream support
External HDMI receivers have analog circuitry that needs to be powered-on
when exiting standby, and a mechanism to detect PCM v. IEC61937 data.
These two steps take time and up to 2-3 seconds of audio may be muted
when starting playback.

Intel hardware (Haswell and beyond) can keep the link active
with a 'silent stream', so that the receiver does not go through those
two steps when valid audio is transmitted. This mechanism relies
on an setting the channel_id as 0xf, sending info packet and preventing
the codec from going to D3,  which will increase the platform
static power consumption. The info packet assumes a basic 2ch stereo,
and the silent stream is enabled when connecting a monitor.
In case of format changes the detection of PCM v. IEC61937 needs to
be re-run. In this case there is no way to avoid the 2-3s mute.

The silent stream is enabled with a Kconfig option, as well as a kernel
parameter should there be a need to override the build time default.
This approach is used based on the power_save capability as an example,
but in the future, it may be used with a kcontrol,
depending on UCM support for HDaudio legacy.

Signed-off-by: Harsha Priya <harshapriya.n@intel.com>
Signed-off-by: Emmanuel Jillela <emmanuel.jillela@intel.com>
Reviewed-by: Kai Vehmanen <kai.vehmanen@linux.intel.com>
Reported-by: kernel test robot <lkp@intel.com>
Link: https://lore.kernel.org/r/1594068797-14011-1-git-send-email-harshapriya.n@intel.com
Signed-off-by: Takashi Iwai <tiwai@suse.de>

(cherry picked from commit 951894cf30f417443d9e0eda33312613c2a7edf6
 https://github.com/tiwai/sound.git master)

Conflicts:
   sound/pci/hda/patch_hdmi.c

BUG=b:153123976
TEST=Play movie with Dell C5518QT monitor

Change-Id: I877f6cc0bcfe9440b20ad76e1465debba89765b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2291207
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Emmanuel Jillela <emmanuel.jillela@intel.corp-partner.google.com>
Commit-Queue: Cheng-Yi Chiang <cychiang@chromium.org>