CHROMIUM: arm64: dts: qcom: sc7180: Adjust display brightness curve
The low end brightness on lazor looked dim. Generally these panels have
a minimum duty cycle they like before producing usable light.

The idea is that we need to crop off the low end of the brightness
curve.

In order to maintain our sanity we bypass userspace's curve (powerd on
chromeos) by rolling the exponential curve in to brightness-levels. This
makes the 'scale' sysfs attribute get set to 'non-linear' so powerd will
yield.

We start with a stretched out pow(x,2) curve that goes to 4096,
then we crop the first few values.

default-brightness-level is required for some reason, I let powerd set
it then copied the value it was set to to the dts. This should assure we
don't have discontinuities when turning on.

BUG=b:155684698
TEST=cat /sys/devices/platform/backlight/backlight/backlight/scale
must show "non-linear"
TEST=Lazor's low end brightness is now usable
TEST=Trogdor is also benefitting from this

Signed-off-by: Alexandru Stan <amstan@chromium.org>
Change-Id: If9cd4bec53885d493e1e185cb132a8fc0dad913d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2291209
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>