FROMGIT: tty: serial: qcom-geni-serial: Drop the icc bw votes in suspend for console
When using the geni-serial as console, its important to be
able to hit the lowest possible power state in suspend,
even with no_console_suspend.
The only thing that prevents it today on platforms like the sc7180
is the interconnect BW votes, which we certainly don't need when
the system is in suspend. So in the suspend handler mark them as
ACTIVE_ONLY (0x3) and on resume switch them back to the ALWAYS tag (0x7)

Signed-off-by: Rajendra Nayak <rnayak@codeaurora.org>
Reviewed-by: Akash Asthana <akashast@codeaurora.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Acked-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Link: https://lore.kernel.org/r/1594704709-26072-1-git-send-email-rnayak@codeaurora.org
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
(cherry picked from commit 4a3107f61f1ce2c8ccd4dde8ae655ae3f2996f35
 git://git.kernel.org/pub/scm/linux/kernel/git/qcom/linux.git drivers-for-5.9)

BUG=b:140573035
TEST=suspend_stress_test

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: I0794fdf8cd5e7ab4c516994fbf4bc6a851b7fb4d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2297226
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>