CHROMIUM: soundwire: SDCA: add helper macro to access controls
The upcoming SDCA (SoundWire Device Class Audio) specification defines
a hiearchical encoding to interface with Class-defined capabilities,
based on which audio function, entity, control and channel being used.

The specification is not yet accessible to the general public but this
information is released with explicit permission from the MIPI Board
to avoid delays with SDCA support on Linux platforms.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Reviewed-by: Rander Wang <rander.wang@linux.intel.com>
Reviewed-by: Guennadi Liakhovetski <guennadi.liakhovetski@linux.intel.com>
Reviewed-by: Bard Liao <yung-chuan.liao@linux.intel.com>

Signed-off-by: Chintan Patel <chintan.m.patel@intel.com>

BUG=b:162008784
TEST=Test audio on volteer.
TOREVERT: Once upstream solution lands revert this.

Change-Id: If03465e8b60a4c8168fd0b7512bc2055134fac5c
Signed-off-by: Mike Mason <michael.w.mason@intel.corp-partner.google.com>