CHROMIUM: dts: qcom: Init backlight to off on sc7180-trogdor family
We don't want the backlight enable pin to go high until we've actually
set the eDP link up.  However, the current Linux PWM backlight driver
sets the backlight enable pin to high in its probe function if it
wasn't already set to be an output.  That means that if you booted up
the firmware in normal mode (you didn't try to show the dev screen and
just left the backlight enable pin as an input) then the backlight
enable pin would be set early.

On Chrome OS devices the firmware never leaves the screen / backlight
on when booting the main OS.  Let's explicitly set the backlight
enable to low in the default pin control so that the backlight driver
doesn't try to set it high.

BUG=b:160560509
TEST=Printouts show that backlight no longer going high in probe.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: I099a9f38fc2d08bd40008ad15bdaee3165fb6c28
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2308590
Reviewed-by: Stephen Boyd <swboyd@chromium.org>