CHROMIUM: drm/mst: Another attempt at fixing port destruction
This patch is yet another attempt at navigating the port refcount
minefield. Some topologies can result in ports destroying other ports
(the most tangible example of this is nested branch devices). So
the previous attempt at adding more serialization resulted in these
topologies to deadlock. So we'll revert back to partial locking in the
port loop and try to make sure sibling ports are all marked for
destruction with valid refcounts at the same time.

Here's a copy of my comment from the code describing my mental model of
what is going on when ports are destroyed and how this latest mitigation
works:

Assume the following topology:
	- branch device 1 		(BD1)
		- port A 		(PA)
		- port B 		(PB)
		- branch device 2 	(BD2)
			- port X 	(PX)
			- port Y 	(PY)

Assume time is measured in T0,T1,T2,etc

Assume at T0 that BD1 is disconnected.

This will cause the drive to realize the port is no longer in MST
mode. The driver calls drm_dp_mst_topology_mgr_set_mst() which will
give up the last reference on BD1 causing a call to
drm_dp_destroy_mst_branch_device() which in turn gives up the final
references on PA and PB.

When the PA reaches 0 refcount at T1, it will add itself to the
destroy_connector_list and schedule this worker. PB will also reach
refcount 0 and do the same at T4.

There are 2 possible times when this worker can run, at T2 or
T5 (ie: before or after PB is marked for destruction).

Case 1 - worker runs at T5, both PA and PB marked for
destruction
  This worker will destroy PA and PB and give up the last references
  on BD2 which will give up the last refs on PX and PY causing this
  worker to be recursively scheduled again to clean up the BD2 ports.

Case 2 - worker runs at T2, only PA marked for destruction
  This worker will destroy PA. On its face, this doesn't seem like a
  problem since the worker would just run again for PB and again for
  BD2. However when PA is destroyed, it updates the payloads on BD1
  and children including PB. If the payload update occurs at T3,
  everything operates as normal. However if the payload update occurs
  at T5 when PB refcount is already 0, we'll end up marking PB for
  destruction twice since the payload update will increase the
  refcount from 0->1 and then back down to 0 (triggering another PB
  port destruction). This is what was happening in
    https://b.corp.google.com/issues/144866969#comment10

Repeat the above scenarios for each branch device in the
tree.

To fix Case 2, we serialized port destroy and this worker such that
port destruction could not interfere with this worker.  Unfortunately
it is not that easy. If BD2 exists, it will lose its last reference
on the worker's stack which causes it to mark PX and PY for
destruction. As mentioned above, port destruction requires the
destroy_connector_lock which will result in deadlock if this worker
is already holding it. This is what's happening in
   https://b.corp.google.com/issues/144866969#comment97 (Case 3)

So for any port we need to:
  1- ensure that all sibling ports have valid refcounts when updating
     payloads
  2- ensure that we can mark ports for destruction without
     serialization between this worker and port destruction

To achieve this, we'll force Case 1 by only scheduling this worker
after all ports for a branch device have been marked for deletion (in
drm_dp_destroy_mst_branch_device()). This ensures that every
invocation of this function we are guaranteed to have all ports for
the top-level branch (BD1) in the destroy_connector_list avoiding
Case 2. This allows us to re-initialize a port's reference count in
drm_dp_destroy_port() so for any port in the destroy list it is also
guaranteed that all sibling ports have valid refcounts.

Further, since BD2 and all other child branch devices are destroyed
on this stack (via drm_dp_port_teardown_pdt()), we are similarly
guaranteed that on each iteration of the below loop all ports from
BD2 and children are in the destroy_connector_list and have valid
reference counts.

BUG=b:144866969
TEST=Tested on eve and atlas with single and chained MST setups

Signed-off-by: Sean Paul <seanpaul@chromium.org>
Change-Id: I03a5429097df37935d6b79a77b67cea7e13c30cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2317505