CHROMIUM: Move drm_dp_put_port() outside of lock in detect()
I don't think this change makes a practical difference if the locking
goes as we expect it should. However if the port is marked for
destruction in detect() we want to make sure this is done outside the
destroy_connector_lock since this will cause a recursive lock.

BUG=b:144866969
TEST=compile tested

Signed-off-by: Sean Paul <seanpaul@chromium.org>
Change-Id: I55e56c964910af7ddf7b3b41c84aacb87f29c121
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2318187
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
(cherry picked from commit 1ba0faa5fab15ad45a956d137b7409ecaf1b2281)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2335332